Create README.md
This is the project which lets the user guess a generated digit code. The program randomly generates a four-digit code and then lets the user try to guess it. If the user guesses all four digits correctly, it congratulates him and prompts him if he wants to play again. If he presses "y" on the keyboard, the program starts again with a new code, and if "N" the program quits.

If the user gets some digits correct, the program gives him feedback on the number of correct digits and a number of misplaced digits then prompts him to enter the code again. If he reaches the maximum number of guesses before getting a correct guess, the program ends and gives a message that he/she has reached a maximum number of guesses then asks him/her if he/she wants to play again.

The program has three classes, project1.h class, project1 class, and main class. The project1.h controls constants declaration such as the length of the code to be generated which is four in our case, the code span which is the boundaries of the code which is from 1 to 6 in our case, and the maximum number of guesses which is 12 in our case. The class also deals with the declaration of functions. This makes the program clear and easy to understand, makes the testing of functions for errors easy, saves time from typing the same functions again and again, and helps in modification of the program without changing its structure.

Project1 class interacts with the user by first giving the introduction about the game and how to play it and then take the user’s input and gives the feedback after comparing the input to the randomly generated code. The first function in this class is “get_player_code()” from line 13 to 88; this prompts the user to enter a four-digit code, then checks the code length and whether the digits are in bounds. Initially, the code is filled with zeros, then filled with whatever input got from the user. Using Boolean which is “true” or “false” statement, valid code is initially set to zero. After that there is a loop created which allows the user to enter the code again whenever he put the wrong code. Lines 38 and 39 prompts the user to enter the code then takes the code entered. Lines 43 to 47 checks if the code has the exact number of digits, in our case four. If yes, the system proceeds to the next step and if no, it gives the error message that the code is out of bound and remind the user that the code should have 4 digits then starts the loop again by prompting the user to reenter the digits. If the number of digits is correct, the system checks each digit and determine if there is any digit out of range or a non-integer. If there is, the system gives the error message that the code contains digits out of range or non-integer then starts the loop from the beginning. 
The random code is generated in the function “generate_code()”, the code is initially set to zero then each zero is replaced by a digit within the range every time the user plays the game. After the code is generated it is compared to check if it matches the one guessed by the user.
The main class interacts with the code entered from the user and compares it to the one generated by the system. The class starts by defining the constants; here you can change the code span to whatever range you want, for this code the number is set not to exceed 6. You can also change the code length which is 4 in our case to whatever number of digits you want. You can also change the number of maximum guesses you want from the user.
After defining the constants, the system then copies the code entered by the user and the code generated by the system for comparison. It creates the counters for correct digits, misplaced digits and player’s guess. The system then creates a loop for the game as long as the code is not guessed correct or the maximum number of guesses is not reached. It starts by creating the space between the user guess, the correct digit, and the misplaced digits to avoid confusion to the user in the time of output. Every time the user enters the four digits code, the system adds one to the number of guesses which was initially set to zero. The system then starts comparing the guessed code and the actual code for output; If the player code is the same as the secret code, the system outputs a congratulation message and tells him or her the number of guesses taken for him/her to guess the right digits then the system breaks there and does not proceed to read the rest of the code. 
In every guess, the system compares each individual digit if it matches the corresponding digit in the secret code. The system increments the number of correct guesses which was initially set to zero by then outputs the number of correct digits by 1. The system also checks if there is a correct digit which is in a wrong position then increments the number of misplaced digits which was initially set to zero by 1 every time there is one. The number of correct digits and misplaced digits are outputted to the user every time he tries to guess the code.
Once the number of guesses exceeds the limit, the system tells the user that, he/she has run out of guesses and then tells him what the right code was. The system then asks the user if he/she would like to play again. If the user writes “N” which means no, the system thanks the user for playing then ten terminates the game. If the user types any letter other than “N”, the system assumes that it is a yes and starts the game again with a new code.